Migrate Collapsible,Searchable,Sortable to Traits
