+++ src/Views/Traits/Configuration/ColumnConfiguration.php
@@ -23,27 +23,10 @@ trait ColumnConfiguration
         return $this;
     }
 
-    public function sortable(?callable $callback = null): self
-    {
-        $this->sortable = true;
-
-        $this->sortCallback = $callback;
-
-        return $this;
-    }
 
     public function format(callable $callable): Column
     {
         $this->formatCallback = $callable;
-
-        return $this;
-    }
-
-    public function searchable(?callable $callback = null): self
-    {
-        $this->searchable = true;
-
-        $this->searchCallback = $callback;
 
         return $this;
     }
@@ -62,20 +45,6 @@ trait ColumnConfiguration
         return $this;
     }
 
-    public function setSortingPillTitle(string $title): self
-    {
-        $this->sortingPillTitle = $title;
-
-        return $this;
-    }
-
-    public function setSortingPillDirections(string $asc, string $desc): self
-    {
-        $this->sortingPillDirectionAsc = $asc;
-        $this->sortingPillDirectionDesc = $desc;
-
-        return $this;
-    }
 
     public function eagerLoadRelations(): self
     {
