+++ src/Views/Traits/Helpers/ColumnHelpers.php
@@ -140,120 +140,6 @@ trait ColumnHelpers
         return $row->{$this->getRelationString().'.'.$this->getField()};
     }
 
-    public function getSortCallback(): ?callable
-    {
-        return $this->sortCallback;
-    }
-
-    public function isSortable(): bool
-    {
-        return $this->hasField() && $this->sortable === true;
-    }
-
-    public function hasSortCallback(): bool
-    {
-        return $this->sortCallback !== null;
-    }
-
-    public function getSearchCallback(): ?callable
-    {
-        return $this->searchCallback;
-    }
-
-    public function isSearchable(): bool
-    {
-        return $this->hasField() && $this->searchable === true;
-    }
-
-    public function hasSearchCallback(): bool
-    {
-        return $this->searchCallback !== null;
-    }
-
-    public function collapseOnMobile(): self
-    {
-        $this->collapseOnMobile = true;
-
-        return $this;
-    }
-
-    public function shouldCollapseOnMobile(): bool
-    {
-        return $this->collapseOnMobile;
-    }
-
-    /**
-     * @return $this
-     */
-    public function collapseOnTablet(): self
-    {
-        $this->collapseOnTablet = true;
-
-        return $this;
-    }
-
-    public function shouldCollapseOnTablet(): bool
-    {
-        return $this->collapseOnTablet;
-    }
-
-    public function collapseAlways(): self
-    {
-        $this->collapseAlways = true;
-
-        return $this;
-    }
-
-    public function shouldCollapseAlways(): bool
-    {
-        return $this->collapseAlways;
-    }
-
-    public function getSortingPillTitle(): string
-    {
-        if ($this->hasCustomSortingPillTitle()) {
-            return $this->getCustomSortingPillTitle();
-        }
-
-        return $this->getTitle();
-    }
-
-    public function getCustomSortingPillTitle(): ?string
-    {
-        return $this->sortingPillTitle;
-    }
-
-    public function hasCustomSortingPillTitle(): bool
-    {
-        return $this->getCustomSortingPillTitle() !== null;
-    }
-
-    public function hasCustomSortingPillDirections(): bool
-    {
-        return $this->sortingPillDirectionAsc !== null && $this->sortingPillDirectionDesc !== null;
-    }
-
-    public function getCustomSortingPillDirections(string $direction): string
-    {
-        if ($direction === 'asc') {
-            return $this->sortingPillDirectionAsc;
-        }
-
-        if ($direction === 'desc') {
-            return $this->sortingPillDirectionDesc;
-        }
-
-        return __('N/A');
-    }
-
-    public function getSortingPillDirection(DataTableComponent $component, string $direction): string
-    {
-        if ($this->hasCustomSortingPillDirections()) {
-            return $this->getCustomSortingPillDirections($direction);
-        }
-
-        return $direction === 'asc' ? $component->getDefaultSortingLabelAsc() : $component->getDefaultSortingLabelDesc();
-    }
 
     public function eagerLoadRelationsIsEnabled(): bool
     {
